pyproject.toml
[tool.poetry]
name = "virustotal-intelligence"
version = "0" # Will be overridden by __version__ in module.
description = """
Connector for OpenCTI that monitors VirusTotal Live Hunt notifications and records sightings for the matched YARA
rules. Will import YARA rule if not already present. Will import file metadata as File observed-data. Requires
VirusTotal Enterprise subscription.
"""
authors = ["Derek Ditch <dcode@elastic.co>"]

# This plugin allows use of __version__ field in __init__.py
[tool.poetry-version-plugin]
source = "init"

[tool.poetry.scripts]
connector = 'virustotal_intelligence.console:main'

[tool.poetry.dependencies]
python = "^3.9"
plyara = "^2.1.1"
pycti = "^5.1.3"
docopt = "^0.6.2"
vt-py = "^0.7.6"
"ruamel.yaml" = "^0.17.17"
ecs-logging = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = "^21.10b0"
flake8 = "<4"
pytest-flake8 = "^1.0.7"
pytest-isort = "^2.0.0"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest-black = "^0.3.12"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--black --isort --flake8"
testpaths = ["."]
flake8-max-line-length = 89
flake8-ignore = "E203 E266 E501 W503"

[tool.flake8]
# Run with "pflake8"
ignore = "E203,E266,E501,W503"
max-line-length = 89
select = "B,C,E,F,W,T4,B9"

[tool.mypy]
files = ["virustotal_intelligence/", "tests/"]

ignore_missing_imports = true
strict_optional = false
allow_redefinition = true
show_error_context = false
show_column_numbers = true
